a2a/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
google-a2a-adk==0.2.2
h11==0.14.0
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.11.0a2
pydantic_core==2.29.0
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
torch==2.6.0
torchaudio==2.6.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.0
websockets==14.2
RealtimeSTT==0.3.93
opencv-python==4.10.0.84
